refact: 削除ボタンのテストを別ファイルに切り出した
